+++ DataRepository/EntityModels/GeminiCommonModel.cs
@@ -1,0 +1,22 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace DataRepository.EntityModels
+{
+    public class GeminiCommonModel
+    {
+        public class Content
+        {
+            public string role { get; set; }
+            public List<Part> parts { get; set; }
+        }
+
+        public class Part
+        {
+            public string text { get; set; }
+        }
+    }
+}
