+++ DataRepository/EntityModels/GeminiRequestModel.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using static DataRepository.EntityModels.GeminiCommonModel;
 
 namespace DataRepository.EntityModels
 {
@@ -27,26 +28,26 @@ namespace DataRepository.EntityModels
             public List<Content> contents { get; set; }
         }
 
-        public class Content
-        {
-            public string role { get; set; }
-            public List<Part> parts { get; set; }
-        }
+        //public class Content
+        //{
+        //    public string role { get; set; }
+        //    public List<Part> parts { get; set; }
+        //}
 
-        public class Part
-        {
-            public string text { get; set; }
-        }
+        //public class Part
+        //{
+        //    public string text { get; set; }
+        //}
 
-        public class GeminiResponse
-        {
-            public List<Candidate> candidates { get; set; }
-        }
+        //public class GeminiResponse
+        //{
+        //    public List<Candidate> candidates { get; set; }
+        //}
 
-        public class Candidate
-        {
-            public Content content { get; set; }
-        }
+        //public class Candidate
+        //{
+        //    public Content content { get; set; }
+        //}
         public class GeminiConversationRequest
         {
             public List<string> ConversationThread { get; set; }
