+++ DataRepository/EntityModels/GeminiResponseModel.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using static DataRepository.EntityModels.GeminiCommonModel;
 
 namespace DataRepository.EntityModels
 {
@@ -23,6 +24,14 @@ namespace DataRepository.EntityModels
             public bool Success { get; set; }
             public string? ErrorMessage { get; set; }
             public string Result { get; set; } = string.Empty;
+        }
+        public class GeminiResponse
+        {
+            public List<Candidate> candidates { get; set; }
+        }
+        public class Candidate
+        {
+            public Content content { get; set; }
         }
     }
 }
